Create unit tests setup, move transformation code to seperate file.
