+++ package.json
@@ -38,7 +38,7 @@
     "@iobroker/adapter-dev": "^1.3.0",
     "@iobroker/adapter-react": "^2.0.22",
     "@iobroker/eslint-config": "^1.0.0",
-    "@iobroker/testing": "^5.0.0",
+    "@iobroker/testing": "^5.0.4",
     "@material-ui/core": "^4.12.4",
     "@tsconfig/node16": "^16.1.3",
     "@types/chai": "^5.0.1",
@@ -50,8 +50,8 @@
     "@types/react-dom": "^17.0.26",
     "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^4.0.0",
-    "chai": "^5.1.2",
-    "chai-as-promised": "^8.0.1",
+    "chai": "^4.5.0",
+    "chai-as-promised": "^7.1.2",
     "eslint": "^9.17.0",
     "eslint-plugin-react": "^7.37.3",
     "globals": "^15.14.0",
@@ -60,7 +60,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "sinon": "^19.0.2",
-    "sinon-chai": "^4.0.0",
+    "sinon-chai": "^3.7.0",
     "typescript": "^4.6.4"
   },
   "main": "main.js",
